+++ DutWrapper.TestRunner/AccountTest.cs
@@ -1,6 +1,8 @@
 ﻿using DutWrapper.Model.Enums;
 using Microsoft.VisualStudio.TestTools.UnitTesting;
+using Newtonsoft.Json;
 using System;
+using System.Diagnostics;
 using System.Net.Http;
 
 namespace DutWrapper.TestRunner
@@ -9,6 +11,9 @@ namespace DutWrapper.TestRunner
     public class AccountTest
     {
         public string SESSION_ID = null;
+        public int YEAR = 22;
+        public int SEMESTER = 1; // 1,2, 3 is 2 in summer
+        public Formatting JSON_FORMATTING = Formatting.Indented;
 
         [TestMethod]
         public void TestEntireAccountFunction()
@@ -22,29 +27,60 @@ namespace DutWrapper.TestRunner
 
             SESSION_ID = Account.GenerateNewSessionId().Result;
             if (SESSION_ID == null)
+            {
                 throw new HttpRequestException("Failed while getting new Session ID! This test cannot continue.");
+            }
+            else Debug.WriteLine($"Session ID: { SESSION_ID }");
+            Debug.WriteLine("");
 
+            Debug.WriteLine($"Processing Account.Login...");
             Account.Login(SESSION_ID, data[0], data[1]).Wait();
+            Debug.WriteLine("Completed task.");
+            Debug.WriteLine("");
 
-            var result = Account.IsLoggedIn(SESSION_ID).Result;
-            Console.WriteLine($"IsLoggedIn: {result} (Session ID: {SESSION_ID})");
-            if (result != LoginStatus.LoggedIn)
+            Debug.WriteLine($"Processing Account.IsLoggedIn...");
+            var resultIsLoggedIn = Account.IsLoggedIn(SESSION_ID).Result;
+            Debug.WriteLine($"Result: { JsonConvert.SerializeObject(resultIsLoggedIn, JSON_FORMATTING) }");
+            if (resultIsLoggedIn != LoginStatus.LoggedIn)
+            {
                 throw new HttpRequestException("Failed while logging you in! This test cannot continue.");
+            }
+            Debug.WriteLine("");
 
-            var result3 = Account.GetSubjectScheduleList(SESSION_ID, 22, 1).Result;
-            Console.WriteLine($"Subject schedule count: {(result3 == null ? "(unknown)" : result3.Count.ToString())} (Session ID: {SESSION_ID})");
-            var result4 = Account.GetSubjectFeeList(SESSION_ID, 22, 1).Result;
-            Console.WriteLine($"Subject fee count: {(result4 == null ? "(unknown)" : result4.Count.ToString())} (Session ID: {SESSION_ID})");
-            var result5 = Account.GetAccountInformation(SESSION_ID).Result;
-            Console.WriteLine($"Is account information null: {result5 == null} (Session ID: {SESSION_ID})");
-            var result6 = Account.GetAccountTrainingResult(SESSION_ID).Result;
-            Console.WriteLine($"Is account training result null: {result6 == null} (Session ID: {SESSION_ID})");
+            Debug.WriteLine($"Processing Account.GetSubjectScheduleList...");
+            var resultSubSchedule = Account.GetSubjectScheduleList(SESSION_ID, YEAR, SEMESTER).Result;
+            Debug.WriteLine($"Result: {JsonConvert.SerializeObject(resultSubSchedule, JSON_FORMATTING)}");
+            Debug.WriteLine($"Count: {(resultSubSchedule == null ? "(unknown)" : resultSubSchedule.Count.ToString())}");
+            Debug.WriteLine("");
 
+            Debug.WriteLine($"Processing Account.GetSubjectFeeList...");
+            var resultSubFee = Account.GetSubjectFeeList(SESSION_ID, YEAR, SEMESTER).Result;
+            Debug.WriteLine($"Result: {JsonConvert.SerializeObject(resultSubFee, JSON_FORMATTING)}");
+            Debug.WriteLine($"Count: {(resultSubFee == null ? "(unknown)" : resultSubFee.Count.ToString())}");
+            Debug.WriteLine("");
+
+            Debug.WriteLine($"Processing Account.GetAccountInformation...");
+            var resultAccInfo = Account.GetAccountInformation(SESSION_ID).Result;
+            Debug.WriteLine($"Result: {JsonConvert.SerializeObject(resultAccInfo, JSON_FORMATTING)}");
+            Debug.WriteLine($"Is account information null: {resultAccInfo == null}");
+            Debug.WriteLine("");
+
+            Debug.WriteLine($"Processing Account.GetAccountTrainingResult...");
+            var resultAccTrainStat = Account.GetAccountTrainingResult(SESSION_ID).Result;
+            Debug.WriteLine($"Result: {JsonConvert.SerializeObject(resultAccTrainStat, JSON_FORMATTING)}");
+            Debug.WriteLine($"Is account training result null: {resultAccTrainStat == null}");
+            Debug.WriteLine("");
+
+            Debug.WriteLine($"Processing Account.Logout...");
             Account.Logout(SESSION_ID).Wait();
-            Console.WriteLine($"Logged out (Session ID: {SESSION_ID})");
+            Debug.WriteLine("Completed task.");
+            Debug.WriteLine("");
 
-            var result2 = Account.IsLoggedIn(SESSION_ID).Result;
-            Console.WriteLine($"IsLoggedIn: {result2} (Session ID: {SESSION_ID})");
+            Debug.WriteLine($"Processing Account.IsLoggedIn...");
+            var resultIsLoggedIn2 = Account.IsLoggedIn(SESSION_ID).Result;
+            Debug.WriteLine($"Result: {JsonConvert.SerializeObject(resultIsLoggedIn2, JSON_FORMATTING)}");
+            Debug.WriteLine($"IsLoggedIn: {resultIsLoggedIn2}");
+            Debug.WriteLine("");
         }
     }
 }